move interfaces in folders with implementations
