+++ Markdown/Markdown/MarkDown.cs
@@ -3,25 +3,27 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Markdown.Interfaces;
+using Markdown.MdTag;
+using Markdown.Parser;
 
 namespace Markdown
 {
     public class MarkDown
     {
-        private IParser<Tag> parser;
+        private MdTagParser parser;
 
         public string Render(string rawMarkdown)
         {
             throw new NotImplementedException();
         }
 
-        public MarkDown(IParser<Tag> parser)
+        public MarkDown(IParser<MdTag.Tag> parser)
         {
+            var a = new Tag("qw", "dwq");
             this.parser = new MdTagParser();
         }
 
-        private string CombineTagsInOneString(List<Tag> tags)
+        private string CombineTagsInOneString(List<MdTag.Tag> tags)
         {
             throw new NotImplementedException();
         }
