+++ Markdown/Markdown/Markdown.csproj
@@ -41,12 +41,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Interfaces\IParser.cs" />
-    <Compile Include="Interfaces\ITag.cs" />
+    <Compile Include="Parser\IParser.cs" />
+    <Compile Include="MdTag\ITag.cs" />
     <Compile Include="MarkDown.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tag.cs" />
-    <Compile Include="MdTagParser.cs" />
+    <Compile Include="MdTag\Tag.cs" />
+    <Compile Include="Parser\MdTagParser.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Tests\" />
