+++ Markdown/Markdown/Parser/IParser.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Markdown.Parser
+{
+    public interface IParser<T>
+    {
+        List<T> Parse(string textToParse);
+    }
+}
