+++ Markdown/Markdown/Parser/MdTagParser.cs
@@ -1,11 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Markdown.Interfaces;
+using Markdown.MdTag;
 
-namespace Markdown
+namespace Markdown.Parser
 {
     class MdTagParser: IParser<Tag>
     {
